[test] update account test
