+++ packages/tools/integration-tests/src/projects/graph/testdata/queries/getAccount.ts
@@ -8,19 +8,14 @@ export function getAccountQuery(accountName: string) {
           guard {
             keys
             predicate
-            __typename
           }
-          __typename
         }
         transactions {
           edges {
             node {
               ...CoreTransactionFields
-              __typename
             }
-            __typename
           }
-          __typename
         }
         transfers {
           edges {
@@ -28,50 +23,35 @@ export function getAccountQuery(accountName: string) {
               ...CoreTransferFields
               crossChainTransfer {
                 ...CoreTransferFields
-                __typename
               }
               transaction {
                 pactId
-                __typename
               }
-              __typename
             }
-            __typename
           }
-          __typename
         }
-        __typename
       }
     }
-
     fragment AllAccountFields on ModuleAccount {
       ...CoreAccountFields
       id
       totalBalance
-      __typename
     }
-
     fragment CoreAccountFields on ModuleAccount {
       accountName
       moduleName
-      __typename
     }
-
     fragment CoreChainAccountFields on ChainModuleAccount {
       balance
       chainId
-      __typename
     }
-
     fragment CoreTransactionFields on Transaction {
       chainId
       code
       creationTime
       height
       requestKey
-      __typename
     }
-
     fragment CoreTransferFields on Transfer {
       amount
       chainId
@@ -79,7 +59,6 @@ export function getAccountQuery(accountName: string) {
       height
       requestKey
       receiverAccount
-      __typename
     }`,
     variables: { moduleName: 'coin', accountName: accountName },
     operationName: 'getAccount',
