+++ packages/tools/integration-tests/src/projects/graph/utils/cryptography-utils.ts
@@ -1,0 +1,7 @@
+export function base64Encode(str: string) {
+  return Buffer.from(str, 'binary').toString('base64');
+}
+
+export function base64Decode(str: string) {
+  return Buffer.from(str, 'base64').toString('binary');
+}
